Add/remove people button working
