+++ assets/js/features/Subscriptions/current-subscriptions/EditSubscriptionsModal.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 
 import Modal from "@/components/Modal";
-import { SubscribersSelector, useSubscriptions } from "@/features/Subscriptions";
-import { useCurrentSubscriptionsContext } from "./CurrentSubscriptionsContext";
-import { getSelectedPeopleFromSubscriptions } from "../utils";
+import { useEditSubscriptionsList } from "@/models/notifications";
+import { useSubscriptions, SubscribersSelectorForm } from "@/features/Subscriptions";
+import { getSelectedPeopleFromSubscriptions } from "@/features/Subscriptions/utils";
+import { useCurrentSubscriptionsContext } from "../CurrentSubscriptions";
+import { SubscribersSelectorProvider } from "../SubscribersSelectorContext";
 
 interface Props {
   isModalOpen: boolean;
@@ -11,13 +13,30 @@ interface Props {
 }
 
 export function EditSubscriptionsModal({ isModalOpen, hideModal }: Props) {
-  const { people, projectName, subscriptionList } = useCurrentSubscriptionsContext();
+  const { people, subscriptionList, callback, type } = useCurrentSubscriptionsContext();
+  const [edit, { loading }] = useEditSubscriptionsList();
+
   const alreadySelected = getSelectedPeopleFromSubscriptions(people, subscriptionList.subscriptions!);
   const subscriptionsState = useSubscriptions(people, { alreadySelected });
 
+  const submitForm = (form) => {
+    const selectedPeopleIds = form.fields.people.value!.map((p) => p.id);
+
+    edit({
+      id: subscriptionList.id,
+      subscriberIds: selectedPeopleIds,
+      type: type,
+    }).then(() => {
+      callback();
+      hideModal();
+    });
+  };
+
   return (
-    <Modal title="Edit subscribers" isOpen={isModalOpen} hideModal={hideModal}>
-      <SubscribersSelector state={subscriptionsState} projectName={projectName} />
-    </Modal>
+    <SubscribersSelectorProvider state={subscriptionsState}>
+      <Modal title="Edit subscribers" isOpen={isModalOpen} hideModal={hideModal} minHeight="200px">
+        <SubscribersSelectorForm callback={submitForm} closeForm={hideModal} loading={loading} />
+      </Modal>
+    </SubscribersSelectorProvider>
   );
 }
