+++ assets/js/features/Subscriptions/index.tsx
@@ -1,5 +1,6 @@
-export { SubscribersSelector } from "./selector/SubscribersSelector";
-export { CurrentSubscriptions } from "./current-subscriptions/CurrentSubscriptions";
+export { SubscribersSelector } from "./SubscribersSelector";
+export { SubscribersSelectorForm } from "./selector/SubscribersSelectorForm";
+export { CurrentSubscriptions } from "./CurrentSubscriptions";
 export { useSubscriptions, SubscriptionsState } from "./useSubscriptions";
 export { findNotifiableProjectContributors } from "./utils";
