+++ assets/js/models/notifications.tsx
@@ -4,7 +4,7 @@ import Api from "@/api";
 import { useUnreadNotificationCount } from "@/api/socket";
 
 export type { SubscriptionList, Subscription } from "@/api";
-export { useSubscribeToNotifications, useUnsubscribeFromNotifications } from "@/api";
+export { useSubscribeToNotifications, useUnsubscribeFromNotifications, useEditSubscriptionsList } from "@/api";
 
 export function useUnreadCount() {
   const [unread, setUnread] = React.useState(0);
